+++ src/main/java/racing/strategy/NumberMoreThanFourConditionStrategy.java
@@ -1,4 +1,4 @@
-package racing.domain;
+package racing.strategy;
 
 import racing.util.RandomNumberGenerator;
